Move savepptx into sc_fileio
